requirements.txt
# Core Dependencies
streamlit>=1.28.0
Pillow>=9.0.0
requests>=2.25.0
pandas>=1.3.0
lxml>=4.6.0

# AI/ML Dependencies
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Advanced UI Dependencies
plotly>=5.15.0
streamlit-authenticator>=0.2.0
streamlit-option-menu>=0.3.0
streamlit-aggrid>=0.3.0
streamlit-elements>=0.1.0

# Data Processing and Export
openpyxl>=3.1.0
xlsxwriter>=3.1.0
reportlab>=4.0.0
jinja2>=3.1.0

# Caching and Performance
redis>=4.6.0
diskcache>=5.6.0
joblib>=1.3.0

# Additional Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
rich>=13.0.0
psutil>=5.9.0

# Build Dependencies
pyinstaller>=5.13.0
auto-py-to-exe>=2.40.0
